Changes to be committed: 	modified:   server/controller/admin/BrandsController.js 	modified:   server/controller/admin/DeviceController.js 	modified:   server/controller/admin/MetersController.js 	modified:   server/models/Brands.js 	modified:   server/routes/admin/brandsRouter.js 	modified:   server/routes/index.js

+++ server/models/Brands.js
@@ -28,37 +28,58 @@ Brands.createBrands = async ({ name }) => {
 
 Brands.getOneBrandById = async (id) => {
     try {
-        const brands = await Brands.findByPk(id);
-        return brands;
+        const brand = await Brands.findByPk(id);
+        if (!brand) {
+            throw new Error("Бренд не найден");
+        }
+        return brand;
     } catch (e) {
         console.error(e);
+        return { error: "Бренд не найден", status: 500 };
     }
 };
 
 Brands.getAllBrands = async () => {
     try {
         const brands = await Brands.findAll();
-        return brands;
+        if (brands.length === 0) {
+            throw new Error("Бренды не найдены");
+        }
+        return { brands };
     } catch (e) {
         console.error(e);
+        return { error: e.message, status: 404 };
     }
 };
 
 Brands.deleteBrands = async (id) => {
     try {
-        const brands = await Brands.destroy({ where: { id } });
-        return brands;
+        const brand = await Brands.findByPk(id);
+        if (!brand) {
+            throw new Error("Бренд не найден");
+        }
+        await Brands.destroy({ where: { id } });
+
+        return brand;
     } catch (e) {
         console.error(e);
+        return { error: "Ошибка при удалении бренда", status: 500 };
     }
 };
 
 Brands.updateBrands = async ({ id, name }) => {
     try {
-        const brands = await Brands.update({ name }, { where: { id } });
-        return brands;
+        const brand = await Brands.findByPk(id);
+        if (!brand) {
+            throw new Error("Бренд не найден");
+        }
+        await Brands.update({ name }, { where: { id } });
+        const newBrand = await Brands.findByPk(id);
+
+        return newBrand;
     } catch (e) {
         console.error(e);
+        return { error: "Ошибка редактирования бренда", status: 500 };
     }
 };
